usb-ehci: Ensure frindex writes leave a valid frindex value
RH-Author: Hans de Goede <hdegoede@redhat.com>
Message-id: <1348253864-3050-4-git-send-email-hdegoede@redhat.com>
Patchwork-id: 42180
O-Subject: [RHEL-6.4 qemu-kvm PATCH 03/22] usb-ehci: Ensure frindex writes leave a valid frindex value
Bugzilla: 805172
RH-Acked-by: Uri Lublin <uril@redhat.com>
RH-Acked-by: Arnon Gilboa <agilboa@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>

frindex is a 14 bits counter, so bits 31-14 should always be 0, and
after the commit titled "usb-ehci: frindex always is a 14 bits counter"
we rely on frindex always being a multiple of 8. I've not seen this in
practice, but theoretically a guest can write a value >= 0x4000 or a value
which is not a multiple of 8 value to frindex, this patch ensures that
things will still work when that happens.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
Upstream commit: 8a771f77e2abfb81f1f33a986810c16ecae54ca7
---
 hw/usb-ehci.c | 4 ++++
 1 file changed, 4 insertions(+)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>