usb-redir: Set ep max_packet_size if available
RH-Author: Hans de Goede <hdegoede@redhat.com>
Message-id: <1348253864-3050-17-git-send-email-hdegoede@redhat.com>
Patchwork-id: 42192
O-Subject: [RHEL-6.4 qemu-kvm PATCH 16/22] usb-redir: Set ep max_packet_size if available
Bugzilla: 805172
RH-Acked-by: Uri Lublin <uril@redhat.com>
RH-Acked-by: Arnon Gilboa <agilboa@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>

This is needed for usb-redir to work properly with the xhci emulation.

RHEL-6: Since RHEL-6 qemu does not have per ep info in the usb-core (and no
xhci either), the added if block is empty, but it is needed for a later patch
which in turn is needed for migration compatibility with upstream.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
Upstream commit: 0fde3b7a826aa654612126cdea9832319a34172c
Conflicts: usb-redir.c
---
 usb-redir.c | 4 ++++
 1 file changed, 4 insertions(+)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>