disable SEP on all CPU models
RH-Author: Eduardo Habkost <ehabkost@redhat.com>
Message-id: <1348851023-31907-12-git-send-email-ehabkost@redhat.com>
Patchwork-id: 42505
O-Subject: [RHEL6 qemu-kvm PATCH 11/18] disable SEP on all CPU models
Bugzilla: 745717
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>
RH-Acked-by: Orit Wasserman <owasserm@redhat.com>

Bugzilla: 745717
Upstream status: not applicable

While we don't fix bz#821463 in the kernel, SEP is not being exposed to
the guest, unless execshield is disabled. This patch changes all CPU
models to not expose SEP to reflect reality and not cause problems on
live-migration to RHEL-6.4 after we fix the kernel bug.

For previous discussions about this, see Message-ID:
 <504839E8.5020407@redhat.com
on the virt-devel archives.

Always disabling SEP on -Mrhel6.3.0 would cause problems if and only if:
 - The host has exec-shield=0;
 - It is running a 32-bit guest;
 - It is migrating backwards, from a 6.4 host to 6.3; and
 - Guest userspace tries to make tricks like detecting at runtime
   (not boot time) if SEP is available.

The necessary conditions are so strict that the solution we (Gleb, Orit,
myself) agreed upon is: just disable SEP on -Mrhel6.3.0[1] and document
a known issue with "exec-shield=0" and 32-bit guests on RHEL-6.3.0
hosts.

 [1] Right now we have to disable it on both rhel6.3.0 and rhel6.4.0,
     until we fix the kernel bug (bz#821463)

     When the kernel bug gets fixed, we can add machine-type
     compatibility bit that will enable SEP only on newer machine-types.
     The task of enabling SEP on the rhel6.4.0 machine-type is tracked
     at bz#821741.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
---
 target-i386/cpuid.c | 26 +++++++++++++-------------
 1 file changed, 13 insertions(+), 13 deletions(-)

Signed-off-by: Michal Novotny <minovotn@redhat.com>