doc: correct default NIC to rtl8139
RH-Author: Avi Kivity <avi@redhat.com>
Message-id: <1349282719-25581-1-git-send-email-avi@redhat.com>
Patchwork-id: 42682
O-Subject: [PATCH qemu-kvm RHEL6.4] doc: correct default NIC to e1000
Bugzilla: 833687
RH-Acked-by: Amos Kong <akong@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Pavel Hrdina <phrdina@redhat.com>

Bugzilla: 833687
Upstream: N/A (qemu-kvm is being phased out; not applicable to qemu.git)

The documentation states that the default NIC is e1000, while in reality
we use rtl8139 (for backward compatibility reasons, and to support Windows XP
out of the box).  Correct the documentation to match reality.

Signed-off-by: Avi Kivity <avi@redhat.com>