monitor: allow qapi and old hmp to share the same dispatch table
RH-Author: Luiz Capitulino <lcapitulino@redhat.com>
Message-id: <1349459193-18226-15-git-send-email-lcapitulino@redhat.com>
Patchwork-id: 42762
O-Subject: [RHEL6.4 qemu-kvm PATCH 14/19] monitor: allow qapi and old hmp to share the same dispatch table
Bugzilla: 832458
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

On upstream, qmp and hmp use different dispatch tables: qmp/qapi handlers
are added to qmp-commands.hx and hmp handlers are added to hmp-commands.hx.

On RHEL6 though, qmp/qapi and hmp share the same dispatch table. This
makes it impossible to have a command with the same name in qmp and hmp.

There are two ways to solve that problem:

 1. Backport the table split patches. This is a lot of work (19 patches
    at a minimum, most will conflict). Besides, we're problably going
    to drop qmp-commands.hx soon upstream. This means that if we
    backport the table split patches now, we'll still be in an
    intermediate point in RHEL6

 2. Let the dispatch table be shared by adding a MONITOR_CMD_QMP_ONLY
    flag and changing the monitor code to check for it. Has the
    disadvantage of not being upstream and is somewhat intrusive, but
    the end result is simple anyway

This patch does item 2, it should be simple & easy to revert if we
later find out that doing 1 is better.

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>