fdc: fix DIR register migration
RH-Author: Pavel Hrdina <phrdina@redhat.com>
Message-id: <f80bd0ff8a884b5a2ea28076ad0fdee598b35a7f.1350312602.git.phrdina@redhat.com>
Patchwork-id: 43178
O-Subject: [RHEL-6.4 qemu-kvm PATCH v5 1/2] fdc: fix DIR register migration
Bugzilla: 854474
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Michal Novotny <minovotn@redhat.com>
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>

During migration the DIR register represented in qemu as
media_changed is not migrated to the target guest.

This cause that if you have a floppy disk mounted before you
start the migration, then the FDC returns after migration that
the media is changed. This change ends in IO errors and you
have to remout the floppy disk.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>
---
 hw/fdc.c | 7 ++++---
 1 file changed, 4 insertions(+), 3 deletions(-)

Signed-off-by: Michal Novotny <minovotn@redhat.com>