block: in commit, determine base image from the top image
RH-Author: Jeffrey Cody <jcody@redhat.com>
Message-id: <3aedbef972b4021796b75af5aefdd8a3a9484a40.1350447475.git.jcody@redhat.com>
Patchwork-id: 43293
O-Subject: [RHEL6.4 qemu-kvm PATCH v4 35/35] block: in commit, determine base image from the top image
Bugzilla: 767233
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Eric Blake <eblake@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>

This simplifies some code and error checking, and also fixes a bug.

bdrv_find_backing_image() should only be passed absolute filenames,
or filenames relative to the chain.  In the QMP message handler for
block commit, when looking up the base do so from the determined top
image, so we know it is reachable from top.

Signed-off-by: Jeff Cody <jcody@redhat.com>

Submitted and reviewed upstream, but not accepted yet.
---
 block/commit.c |  7 -------
 blockdev.c     | 21 +++++++++++----------
 2 files changed, 11 insertions(+), 17 deletions(-)

Signed-off-by: Michal Novotny <minovotn@redhat.com>