e1000: Don't set the Capabilities List bit
RH-Author: Michael S. Tsirkin <mst@redhat.com>
Message-id: <d0a70cca9ae928b03ecddff930d54b37c8dd2993.1350572959.git.mst@redhat.com>
Patchwork-id: 43359
O-Subject: [PATCHv2 RHEL6.4 qemu-kvm 3/3] e1000: Don't set the Capabilities List bit
Bugzilla: 866736
RH-Acked-by: Jason Baron <jbaron@redhat.com>
RH-Acked-by: Xiao Wang <jasowang@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

The commit below (appeared in qemu 1.0) removes the capabilities bit,
for spec compliance. We can't remove it unconditionally to avoid
breaking migration, so add a compat flag and keep it on for
machine types 6.3.0 and older.

Author: dann frazier <dann.frazier@canonical.com>

[Originally sent to qemu-kvm list, but I was redirected here]

The Capabilities Pointer is NULL, so this bit shouldn't be set. The state of
this bit doesn't appear to change any behavior on Linux/Windows versions we've
tested, but it does cause Windows' PCI/PCI Express Compliance Test to balk.

I happen to have a physical 82540EM controller, and it also sets the
Capabilities Bit, but it actually has items on the capabilities list to go
with it :)

Signed-off-by: dann frazier <dann.frazier@canonical.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
(cherry-picked from commit dd8e93799f13ef82d83c185b8e71e049452f7d40)
---
 hw/e1000.c | 16 ++++++++++++++--
 hw/pc.c    |  4 ++++
 2 files changed, 18 insertions(+), 2 deletions(-)

Signed-off-by: Michal Novotny <minovotn@redhat.com>