target-i386: kvm: set vcpu_id to APIC ID instead of CPU index
RH-Author: Eduardo Habkost <ehabkost@redhat.com>
Message-id: <1351536756-16475-4-git-send-email-ehabkost@redhat.com>
Patchwork-id: 43761
O-Subject: [RHEL6.4 qemu-kvm PATCH 03/10] target-i386: kvm: set vcpu_id to APIC ID instead of CPU index
Bugzilla: 733720
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>
RH-Acked-by: Gleb Natapov <gleb@redhat.com>

Bugzilla: 733720
Related TestOnly BZs: 816804, 815958, 782648
Upstream status: submitted
 Message-Id: <1351101001-14589-7-git-send-email-ehabkost@redhat.com
 http://article.gmane.org/gmane.comp.emulators.qemu/177690

The CPU ID in KVM is supposed to be the APIC ID, so change the
KVM_CREATE_VCPU call to match it. The current behavior didn't break
anything yet because today the APIC ID is assumed to be == the CPU
index, but this won't be true in the future.

[RHEL note: this patch touches qemu-kvm-x86.c instead of
 target-i386/kvm.c, as the upstream KVM initialization code is very
 different, now]

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
---
 qemu-kvm-x86.c | 2 +-
 1 file changed, 1 insertion(+), 1 deletion(-)

Signed-off-by: Michal Novotny <minovotn@redhat.com>