e1000: Don't set the Capabilities List bit
The commit below (appeared in qemu 1.0) removes the capabilities bit,
for spec compliance. We need to restore it before migration to avoid
breaking migration.

Author: dann frazier <dann.frazier@canonical.com>

[Originally sent to qemu-kvm list, but I was redirected here]

The Capabilities Pointer is NULL, so this bit shouldn't be set. The state of
this bit doesn't appear to change any behavior on Linux/Windows versions we've
tested, but it does cause Windows' PCI/PCI Express Compliance Test to balk.

I happen to have a physical 82540EM controller, and it also sets the
Capabilities Bit, but it actually has items on the capabilities list to go
with it :)

Signed-off-by: dann frazier <dann.frazier@canonical.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
(cherry-picked from commit dd8e93799f13ef82d83c185b8e71e049452f7d40)