qemu-ga: installation docs for win32 platform
The Windows version of the guest agent is being distributed currently
by means of a folder containing a zip file of the the guest agent
and required DLLs.  This adds the README file that will be used in the
zip file, to instruct the user on how to install the guest agent inside
their Windows guest.

Signed-off-by: Jeff Cody <jcody@redhat.com>