cutils: change strtosz_suffix_unit function
if value to be translated is larger than INT64_MAX,
this function will not be convenient for caller to
be aware of it, so change a little for this.

Signed-off-by: liguang <lig.fnst@cn.fujitsu.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
(cherry picked from commit 37edbf7ea8067262a5c3d8bbe4786139348c8311)

RHEL 6: This patch required manual conflict resolution. We're missing
commits ea5b1cfe and a7261486 downstream, and especially the latter
depends on other unrelated patches, so I refrained from cherry-picking
them.

Signed-off-by: Kevin Wolf <kwolf@redhat.com>