dataplane: add host memory mapping code
RH-Author: Stefan Hajnoczi <stefanha@redhat.com>
Message-id: <1357138959-1918-5-git-send-email-stefanha@redhat.com>
Patchwork-id: 45517
O-Subject: [RHEL6.4 qemu-kvm PATCH v5 04/16] dataplane: add host memory mapping code
Bugzilla: 877836
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

The data plane thread needs to map guest physical addresses to host
pointers.  Normally this is done with cpu_physical_memory_map() but the
function assumes the global mutex is held.  The data plane thread does
not touch the global mutex and therefore needs a thread-safe memory
mapping mechanism.

Hostmem registers a CPUPhysMemoryClient similar to how vhost collects
and pushes memory region information into the kernel.  There is a
fine-grained lock on the regions list which is held during lookup and
when updating the regions list.

When the physical memory map changes the .set_memory() callback is
invoked.  It updates the regions list in-place by removing, merging,
splitting, or adding regions.

This patch implements Hostmem for RHEL differently from upstream.  Since
MemoryListener and the memory API is not available in RHEL we use the
vhost's CPUPhysMemoryClient approach.

Note that the VGA region is skipped since dirty logging is not
implemented by Hostmem and we don't expect to DMA into VGA memory.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
---
 Makefile.target        |   3 +-
 configure              |   1 +
 hw/dataplane/hostmem.c | 124 +++++++++++++++++++++++++++++++++++++++++++++++++
 hw/dataplane/hostmem.h |  41 ++++++++++++++++
 4 files changed, 168 insertions(+), 1 deletion(-)
 create mode 100644 hw/dataplane/hostmem.c
 create mode 100644 hw/dataplane/hostmem.h

Signed-off-by: Michal Novotny <minovotn@redhat.com>