dataplane: add virtqueue vring code
RH-Author: Stefan Hajnoczi <stefanha@redhat.com>
Message-id: <1357138959-1918-6-git-send-email-stefanha@redhat.com>
Patchwork-id: 45518
O-Subject: [RHEL6.4 qemu-kvm PATCH v5 05/16] dataplane: add virtqueue vring code
Bugzilla: 877836
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Asias He <asias@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

The virtio-blk-data-plane cannot access memory using the usual QEMU
functions since it executes outside the global mutex and the memory APIs
are this time are not thread-safe.

This patch introduces a virtqueue module based on the kernel's vhost
vring code.  The trick is that we map guest memory ahead of time and
access it cheaply outside the global mutex.

Once the hardware emulation code can execute outside the global mutex it
will be possible to drop this code.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
---
 Makefile.target      |   2 +-
 hw/dataplane/vring.c | 368 +++++++++++++++++++++++++++++++++++++++++++++++++++
 hw/dataplane/vring.h |  62 +++++++++
 trace-events         |   3 +
 4 files changed, 434 insertions(+), 1 deletion(-)
 create mode 100644 hw/dataplane/vring.c
 create mode 100644 hw/dataplane/vring.h

Signed-off-by: Michal Novotny <minovotn@redhat.com>