dataplane: add event loop
RH-Author: Stefan Hajnoczi <stefanha@redhat.com>
Message-id: <1357138959-1918-8-git-send-email-stefanha@redhat.com>
Patchwork-id: 45520
O-Subject: [RHEL6.4 qemu-kvm PATCH v5 07/16] dataplane: add event loop
Bugzilla: 877836
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

Outside the safety of the global mutex we need to poll on file
descriptors.  I found epoll(2) is a convenient way to do that, although
other options could replace this module in the future (such as an
AioContext-based loop or glib's GMainLoop).

One important feature of this small event loop implementation is that
the loop can be terminated in a thread-safe way.  This allows QEMU to
stop the data plane thread cleanly.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
---
 Makefile.target           |   1 +
 hw/dataplane/event-poll.c | 100 ++++++++++++++++++++++++++++++++++++++++++++++
 hw/dataplane/event-poll.h |  40 +++++++++++++++++++
 3 files changed, 141 insertions(+)
 create mode 100644 hw/dataplane/event-poll.c
 create mode 100644 hw/dataplane/event-poll.h

Signed-off-by: Michal Novotny <minovotn@redhat.com>