virtio-blk: add x-data-plane=on|off performance feature
RH-Author: Stefan Hajnoczi <stefanha@redhat.com>
Message-id: <1357138959-1918-16-git-send-email-stefanha@redhat.com>
Patchwork-id: 45528
O-Subject: [RHEL6.4 qemu-kvm PATCH v5 15/16] virtio-blk: add x-data-plane=on|off performance feature
Bugzilla: 877836
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Asias He <asias@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

The virtio-blk-data-plane feature is easy to integrate into
hw/virtio-blk.c.  The data plane can be started and stopped similar to
vhost-net.

Users can take advantage of the virtio-blk-data-plane feature using the
new -device virtio-blk-pci,x-data-plane=on property.

The x-data-plane name was chosen because at this stage the feature is
experimental and likely to see changes in the future.

If the VM configuration does not support virtio-blk-data-plane an error
message is printed.  Although we could fall back to regular virtio-blk,
I prefer the explicit approach since it prompts the user to fix their
configuration if they want the performance benefit of
virtio-blk-data-plane.

Limitations:
 * Only format=raw is supported
 * Live migration is not supported
 * Block jobs, hot unplug, and other operations fail with -EBUSY
 * I/O throttling limits are ignored
 * Only Linux hosts are supported due to Linux AIO usage

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
---
 hw/virtio-blk.c | 54 ++++++++++++++++++++++++++++++++++++++++++++++++++++++
 hw/virtio-pci.c |  4 ++++
 2 files changed, 58 insertions(+)

Signed-off-by: Michal Novotny <minovotn@redhat.com>