qxl: stop using non revision 4 rom fields for revision < 4
RH-Author: Alon Levy <alevy@redhat.com>
Message-id: <1358876003-30877-2-git-send-email-alevy@redhat.com>
Patchwork-id: 47627
O-Subject: [PATCHv2 RHEL-6.4 qemu-kvm 1/2] qxl: stop using non revision 4 rom fields for revision < 4
Bugzilla: 869981
RH-Acked-by: Hans de Goede <hdegoede@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>

interface_set_client_capabilities is only usable if the revision is >=4,
when both a new interrupt notifies of any changes and a new field in the
ROM contains the capabilities. A guest would not expect this interrupt
if it knows the device is revision <= 3. Instead of issueing the
interrupt ignore the API call and log it by a new trace event.

Signed-off-by: Alon Levy <alevy@redhat.com>

upstream: acked but not committed.
---
 hw/qxl.c     | 6 ++++++
 trace-events | 1 +
 2 files changed, 7 insertions(+)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>