qxl: change rom size to 8192
RH-Author: Alon Levy <alevy@redhat.com>
Message-id: <1358876003-30877-3-git-send-email-alevy@redhat.com>
Patchwork-id: 47626
O-Subject: [PATCHv2 RHEL-6.4 qemu-kvm 2/2] qxl: change rom size to 8192
Bugzilla: 869981
RH-Acked-by: Hans de Goede <hdegoede@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>

This is a simpler solution to RHBZ#869981, where migration breaks since
qxl's rom bar size has changed. Instead of ignoring fields in QXLRom,
which is what has actually changed, we remove some of the modes, a
mechanism already accounted for by the guest. The modes left allow for
portrait and landscape only modes, corresponding to orientations 0 and
1.  Orientations 2 and 3 are dropped.

Added assert so that rom size will fit the future QXLRom increases via
spice-protocol changes.

This patch has been tested with 6.1.0.10015. With the newer 6.1.0.10016
there are problems with both "(flipped)" modes prior to the patch, and
the patch loses the ability to set "Portrait" modes. But this is a
separate bug to be fixed in the driver, and besides the patch doesn't
affect the new arbitrary mode setting functionality.

Signed-off-by: Alon Levy <alevy@redhat.com>

upstream: acked but not committed.

RHBZ: 869981
---
 hw/qxl.c | 16 +++++++++-------
 1 file changed, 9 insertions(+), 7 deletions(-)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>