redhat:spec.template: qemu-kvm*-rhev: don't conflict with itself
The current qemu-kvm-rhel package (and its subpackages) has a problem:
it conflicts with itself, confusing "rpm -V" and maybe other tools.

This patch fixes this problem by making the RHEL package have a
"qemu-kvm-rhel" Provides, and by making qemu-kvm-rhev conflict with
qemu-kvm-rhel (any version), and with older versions of qemu-kvm that
didn't have the qemu-kvm-rhel Provides.

The behavior described is reproduced on all subpackages, using a macro
to make sure the rules follow the same pattern on all subpackages.

signed-off-by: Eduardo Habkost <ehabkost@redhat.com>