script: git script to compile every commit in a range of commits
RH-Author: Jeffrey Cody <jcody@redhat.com>
Message-id: <92be7e83e07fedf18e6e7c0aa44526f51a29b4d5.1369944427.git.jcody@redhat.com>
Patchwork-id: 51698
O-Subject: [RHEL6.5 qemu-kvm PATCH v3 1/2] script: git script to compile every commit in a range of commits
Bugzilla:

This is a git script that will iterate through every commit in a
specified range, and perform a configure and make.  The intention
of this script is not to act as a check of code correctness, but to see
if any commit breaks compilation of the tree.

The range passed in to the script is in the form of a normal git range;
the default is HEAD^! (i.e. just the latest commit).

Options may be passed in for configure, as well as make.  The log output
filename and directory may also be optionally specified.

If everything compiled without error, then the exit code from the
program will be 0.

Signed-off-by: Jeff Cody <jcody@redhat.com>
---
 redhat/scripts/git-compile-check | 199 +++++++++++++++++++++++++++++++++++++++
 1 file changed, 199 insertions(+)
 create mode 100755 redhat/scripts/git-compile-check

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>