script: git script to compare backported patches against upstream
RH-Author: Jeffrey Cody <jcody@redhat.com>
Message-id: <98a21d7f5375877b5ccfe1a643c83a2bd2a59608.1369944427.git.jcody@redhat.com>
Patchwork-id: 51699
O-Subject: [RHEL6.5 qemu-kvm PATCH v3 2/2] script: git script to compare backported patches against upstream
Bugzilla:
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

When backporting patches into RHEL from upstream qemu, if we are
lucky the patches are clean cherry-picks, with no conflicts or
differences from upstream.  However, often there are (unfortunately)
differences from upstream.  And, on rare occasions, there are
RHEL-only patches.

This script is meant to act as an aid to reviewers of backports.
Once a series of patches have been applied to a branch, you
can run the script to provide easy-to-read output. It shows
which patches are functionally identical, which patches differ
(and by roughly how much), and which patches are downstream only.

For patches that differ from upstream, it optionally will run a diff
tool (default is 'meld') to show the diff of the diff.

This script assumes your git repo has remotes added for both downstream,
and upstream repositories.

Signed-off-by: Jeff Cody <jcody@redhat.com>
---
 redhat/scripts/git-backport-diff | 302 +++++++++++++++++++++++++++++++++++++++
 1 file changed, 302 insertions(+)
 create mode 100755 redhat/scripts/git-backport-diff

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>