qapi: Introduce blockdev-group-snapshot-sync command
RH-Author: Fam Zheng <famz@redhat.com>
Message-id: <1371795611-7208-10-git-send-email-famz@redhat.com>
Patchwork-id: 52083
O-Subject: [PATCH RHEL-6.5 qemu-kvm v3 09/15] qapi: Introduce blockdev-group-snapshot-sync command
Bugzilla: 956825
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

From: Jeff Cody <jcody@redhat.com>

This is a QAPI/QMP only command to take a snapshot of a group of
devices. This is similar to the blockdev-snapshot-sync command, except
blockdev-group-snapshot-sync accepts a list devices, filenames, and
formats.

It is attempted to keep the snapshot of the group atomic; if the
creation or open of any of the new snapshots fails, then all of
the new snapshots are abandoned, and the name of the snapshot image
that failed is returned.  The failure case should not interrupt
any operations.

Rather than use bdrv_close() along with a subsequent bdrv_open() to
perform the pivot, the original image is never closed and the new
image is placed 'in front' of the original image via manipulation
of the BlockDriverState fields.  Thus, once the new snapshot image
has been successfully created, there are no more failure points
before pivoting to the new snapshot.

This allows the group of disks to remain consistent with each other,
even across snapshot failures.

Signed-off-by: Jeff Cody <jcody@redhat.com>
Acked-by: Luiz Capitulino <lcapitulino@redhat.com>
Signed-off-by: Kevin Wolf <kwolf@redhat.com>
(cherry picked from commit 8802d1fdd4b73e02ce13fb3a233c64c1913634ab)

Backport again for io limit fileds in bdrv_append, since previous
backport (c032ff2a) of this commit dropped changes.

Signed-off-by: Fam Zheng <famz@redhat.com>
---
 block.c | 22 ++++++++++++++++++++++
 1 file changed, 22 insertions(+)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>