qapi: untangle next_list
RH-Author: Luiz Capitulino <lcapitulino@redhat.com>
Message-id: <1373565275-8756-4-git-send-email-lcapitulino@redhat.com>
Patchwork-id: 52445
O-Subject: [RHEL6.5 qemu-kvm PATCH 3/4] qapi: untangle next_list
Bugzilla: 983635
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

From: Paolo Bonzini <pbonzini@redhat.com>

Right now, the semantics of next_list are complicated.  The caller must:

* call start_list

* call next_list for each element *including the first*

* on the first call to next_list, the second argument should point to
NULL and the result is the head of the list.  On subsequent calls,
the second argument should point to the last node (last result of
next_list) and next_list itself tacks the element at the tail of the
list.

This works for both input and output visitor, but having the visitor
write memory when it is only reading the list is ugly.  Plus, relying
on *list to detect the first call is tricky and undocumented.

We can initialize so->entry in next_list instead of start_list, leaving
it NULL in start_list.  This way next_list sees clearly whether it is
on the first call---as a bonus, it discriminates the cases based on
internal state of the visitor rather than external state.  We can
also pull the assignment of the list head from generated code up to
next_list.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Reviewed-by: Michael Roth <mdroth@linux.vnet.ibm.com>
Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
(cherry picked from commit 3a86a0fa76b5103a122b6e817b3827b2837f4956)

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>

Conflicts:
	qapi/qmp-input-visitor.c

[RHEL6 NOTE: qemu_malloc/g_malloc conflict]
---
 docs/qapi-code-gen.txt   |  4 ++--
 qapi/qmp-input-visitor.c | 22 +++++++++++++---------
 scripts/qapi-visit.py    |  4 ++--
 3 files changed, 17 insertions(+), 13 deletions(-)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>