rbd: link and load librbd dynamically
RH-Author: Kevin Wolf <kwolf@redhat.com>
Message-id: <1376470370-32441-3-git-send-email-kwolf@redhat.com>
Patchwork-id: 53330
O-Subject: [RHEL-6.5 qemu-kvm PATCH 2/6] rbd: link and load librbd dynamically
Bugzilla: 988079
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Jeffrey Cody <jcody@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

Bugzilla: 988079
Upstream status: Rejected

This is the downstream-only part that gets us rid of the build-time
dependency on librbd and loads it dynamically when using an image.
It is based on a patch submitted to qemu-devel and archived as
http://lists.gnu.org/archive/html/qemu-devel/2013-04/msg01814.html

Original commit message follows:

This allows the rbd block driver to detect symbols in the installed
version of librbd, and enable or disable features appropriately.  This
obviates the #ifdefs regarding librbd versions.

Loading librbd dynamically also makes the rbd block driver easier to
install and package, since it removes the dependency on librbd at
build time.

Add structures containing the necessary function pointer signatures
and types from librbd, and fill them in the first time the rbd module
is used. Use glib's g_module interface so we don't preclude future
portability, and don't have to deal with odd dlopen behavior directly.

Internally, librbd and some libraries it depends on use C++ templates,
which mean that they each contain a defined weak symbol for their
definition.  Due to the way the linker resolves duplicate symbols, the
libraries loaded by librbd end up using the template definitions from
librbd, creating a circular dependency. This means that once librbd is
loaded, it won't be unloaded. Changing this behavior might work with a
Sun ld, but there doesn't seem to be a portable (or even working with
GNU ld) way to hide these C++ symbols correctly. Instead, never unload
librbd, and explicitly make it resident.

Signed-off-by: Josh Durgin <josh.durgin@inktank.com>
Signed-off-by: Kevin Wolf <kwolf@redhat.com>
---
 Makefile.objs     |   2 +-
 block/rbd.c       | 304 +++++++++++++++++++++++++++++++++++++-----------------
 block/rbd_types.h |  95 +++++++++++++++++
 configure         |  58 +----------
 4 files changed, 311 insertions(+), 148 deletions(-)
 create mode 100644 block/rbd_types.h

Signed-off-by: Michal Novotny <minovotn@redhat.com>