rbd: Only look for qemu-specific copy of librbd.so.1
RH-Author: Kevin Wolf <kwolf@redhat.com>
Message-id: <1376470370-32441-4-git-send-email-kwolf@redhat.com>
Patchwork-id: 53331
O-Subject: [RHEL-6.5 qemu-kvm PATCH 3/6] rbd: Only look for qemu-specific copy of librbd.so.1
Bugzilla: 988079
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Jeffrey Cody <jcody@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

Bugzilla: 988079
Upstream status: n/a

For non-technical reasons, we don't want to pick up a random librbd that
may be installed for other reasons on this system, but only a librbd
that was specificially installed in order to consciously extend qemu.

So change the gmodule_open() argument to an absolute path.

Signed-off-by: Kevin Wolf <kwolf@redhat.com>
---
 block/rbd.c | 2 +-
 1 file changed, 1 insertion(+), 1 deletion(-)

Signed-off-by: Michal Novotny <minovotn@redhat.com>