block: call the snapshot handlers of the protocol drivers
RH-Author: Kevin Wolf <kwolf@redhat.com>
Message-id: <1376470370-32441-7-git-send-email-kwolf@redhat.com>
Patchwork-id: 53334
O-Subject: [RHEL-6.5 qemu-kvm PATCH 6/6] block: call the snapshot handlers of the protocol drivers
Bugzilla: 988079
RH-Acked-by: Fam Zheng <famz@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Jeffrey Cody <jcody@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

From: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>

Bugzilla: 988079

When snapshot handlers are not defined in the format driver, it is
better to call the ones of the protocol driver.  This enables us to
implement snapshot support in the protocol driver.

We need to call bdrv_close() and bdrv_open() handlers of the format
driver before and after bdrv_snapshot_goto() call of the protocol.  It is
because the contents of the block driver state may need to be changed
after loading vmstate.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Kevin Wolf <kwolf@redhat.com>
(cherry picked from commit 7cdb1f6d305e1000b5f882257cbee71b8bb08ef5)

Signed-off-by: Kevin Wolf <kwolf@redhat.com>
---
 block.c | 61 +++++++++++++++++++++++++++++++++++++++++++------------------
 1 file changed, 43 insertions(+), 18 deletions(-)

Signed-off-by: Michal Novotny <minovotn@redhat.com>