add timestamp to error_report()
RH-Author: Laszlo Ersek <lersek@redhat.com>
Message-id: <1376473582-31743-2-git-send-email-lersek@redhat.com>
Patchwork-id: 53361
O-Subject: [RHEL-6.5 qemu-kvm PATCH 1/2] add timestamp to error_report()
Bugzilla: 906931
RH-Acked-by: Michal Novotny <minovotn@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Luiz Capitulino <lcapitulino@redhat.com>

[Issue]
When we offer a customer support service and a problem happens
in a customer's system, we try to understand the problem by
comparing what the customer reports with message logs of the
customer's system.

In this case, we often need to know when the problem happens.

But, currently, there is no timestamp in qemu's error messages.
Therefore, we may not be able to understand the problem based on
error messages.

[Solution]
Add a timestamp to qemu's error message logged by
error_report() with g_time_val_to_iso8601().

Signed-off-by: Seiji Aguchi <seiji.aguchi@hds.com>
Reviewed-by: Stefan Hajnoczi <stefanha@redhat.com>
Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
(cherry picked from commit 5e2ac5191772dea782ff78e95edd395985273019)

Conflicts:

	qemu-error.h
	qemu-options.hx
	vl.c

RHEL-6 notes:
- The upstream "vl.c" hunks have been distributed between RHEL-6 "vl.c"
  and "qemu-config.c", adjacently to / following the pattern of the
  "qemu_realtime_opts" backport (RHEL-6 commit c4dd58ce).
- Other hunks are subject to file renames.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>
---
 qemu-error.h    |    3 +++
 qemu-config.c   |   13 +++++++++++++
 qemu-error.c    |   10 ++++++++++
 vl.c            |   13 +++++++++++++
 qemu-options.hx |   10 ++++++++++
 5 files changed, 49 insertions(+), 0 deletions(-)

Signed-off-by: Michal Novotny <minovotn@redhat.com>