qemu-iotests: Fixed test case 026
Message-id: <1379499002-5231-5-git-send-email-mreitz@redhat.com>
Patchwork-id: 54435
O-Subject: [RHEL-7.0 qemu-kvm PATCH 4/6] qemu-iotests: Fixed test case 026
Bugzilla: 1006959
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Fam Zheng <famz@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>

BZ: 1006959

The reference output for test case 026 hasn't been updated in a long
time and it's one of the "known failing" cases. This patch updates the
reference output so that unintentional changes can be reliably detected
again.

The problem with this test case is that it produces different output
depending on whether -nocache is used or not. The solution of this patch
is to actually have two different reference outputs. If nnn.out.nocache
exists, it is used as the reference output for -nocache; otherwise,
nnn.out stays valid for both cases.

Signed-off-by: Kevin Wolf <kwolf@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
(cherry picked from commit 8f94b077877151de93a63c73f796897309568ddb)

Signed-off-by: Max Reitz <mreitz@redhat.com>